Add client test, fix typos
