+++ tests/US_Enrichment/ClientTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace SmartyStreets\PhpSdk\Tests\US_Enrichment;
+
+require_once(dirname(dirname(__FILE__)) . '/Mocks/MockSerializer.php');
+require_once(dirname(dirname(__FILE__)) . '/Mocks/MockDeserializer.php');
+require_once(dirname(dirname(__FILE__)) . '/Mocks/RequestCapturingSender.php');
+require_once(dirname(dirname(__FILE__)) . '/Mocks/MockStatusCodeSender.php');
+require_once(dirname(dirname(__FILE__)) . '/Mocks/MockSender.php');
+require_once(dirname(dirname(__FILE__)) . '/Mocks/MockCrashingSender.php');
+require_once(dirname(dirname(dirname(__FILE__))) . '/src/US_Enrichment/Client.php');
+require_once(dirname(dirname(dirname(__FILE__))) . '/src/URLPrefixSender.php');
+use SmartyStreets\PhpSdk\Tests\Mocks\MockSerializer;
+use SmartyStreets\PhpSdk\Tests\Mocks\RequestCapturingSender;
+use SmartyStreets\PhpSdk\URLPrefixSender;
+use SmartyStreets\PhpSdk\US_Enrichment\Client;
+use PHPUnit\Framework\TestCase;
+
+class ClientTest extends TestCase {
+
+    public function testSendingLookup() {
+        $capturingSender = new RequestCapturingSender();
+        $sender = new URLPrefixSender("http://localhost/", $capturingSender);
+        $serializer = new MockSerializer(null);
+        $client = new Client($sender, $serializer);
+        $smartyKey = "123";
+        $client->sendPropertyPrincipalLookup($smartyKey);
+
+        $this->assertEquals("http://localhost/123/property/principal?", $capturingSender->getRequest()->getUrl());
+    }
+}
